s/resource/entity/g
On discussion, and based on attempting to explain concepts of Apigility to
users, the ZF team realized we had some naming issues in some of the libraries.

We have 3 concepts:

- entities, which are individual objects with identifiers that we wish to
  address via the URI, and which may contain additional entities or collections
  of entities.
- collections, which contain sets of entities that are of the same type.
- resources, which refer to a collection and the entities within a collection.

In zf-hal, we had a "Resource" object, which was a mis-nomer; these were
actually describing entities. This commit renames resources to entities where
the actual item being referred to should be an entity, and marks a number of
methods and events as deprecated.