Update Hal Plugin to embed entities with $id = 0
Antoher change related to issue https://github.com/zfcampus/zf-hal/issues/44 .
The entity with `$id = 0` was also not properly embedded in the collection because of another similar faulty if clause, this time in the `extractCollection` method. This change fixes the issue and makes sure that the entity is properly rendered.