Generalize marshalSelfLinkFromMetadata
We ended up adding some custom functionality to our hal plugin by extending the existing `Hal.php`. We also use the metadata to marshal links of other links than 'self' links. Would it be possible to generalize the marshalSelfLinkFromMetadata to marshalLinkFromMetadata with as default value for $relation 'self'. This allows users to use the method for rendering other links with the same method.