Allow usage with zend-hydrator v3
There are two key differences in zend-hydrator v3 we need to
accommodate.

First, we have added the `HydratorPluginMangerInterface` to
zend-hydrator v3. This means that any references to
`HydratorPluginManager` when doing type-hinting need to be expanded to
allow the more general interface. This patch updates any cases where
`HydratorPluginManager` was accepted as an argument to make the argument
mixed, and do type-checking internally; cases where
`HydratorPluginManager` was returned have been updated to allow either
type.

Second, the `ExtractionInterface` was updated in v3 to add parameter and
return typehints, which affects the `EntityExtractor`. To accommodate
the two different signatures, this patch creates two versions of that
class, and then adds an autoloading rule that aliases the original name
to the correct version.